[refactor] devide create delete done Todo function from main
