+++ src/js/utils.js
@@ -29,3 +29,18 @@ export function isStatusDeleted(todoItem) {
 export function isStatusTodo(todoItem) {
     return todoItem?.status === statusEnum.TODO;
 }
+
+/**
+* @param {string} inputContent todo content field.
+* @param {statusEnum} inputStatus todo status filld.
+*/
+export function appendTodos(todos, inputContent, inputStatus) {
+   /**@type {Todo}*/
+   const newTask = {
+       id: crypto.randomUUID(),
+       content: inputContent,
+       status: inputStatus
+   }
+   todos.push(newTask);
+}
+
